service-1-llm-service/requirements.txt
fastapi==0.115.11
openai==1.66.3
pydantic==2.10.6
python-dotenv==1.0.1
uvicorn==0.34.0
requests==2.32.3
pytest==8.3.5
